+++ modelstore/models/lightgbm.py
@@ -42,7 +42,7 @@ class LightGbmManager(ModelManager):
         return {"library": "lightgbm", "type": type(kwargs["model"]).__name__}
 
     def _model_data(self, **kwargs) -> dict:
-        """ Returns meta-data about features used to train the model """
+        """ Returns meta-data about the data used to train the model """
         return {}
 
     def _get_functions(self, **kwargs) -> list: