+++ modelstore/models/modelmanager.py
@@ -90,7 +90,7 @@ class ModelManager(ABC):
 
     @abstractmethod
     def _model_data(self, **kwargs) -> dict:
-        """ Returns meta-data about features used to train the model """
+        """ Returns meta-data about the data used to train the model """
         raise NotImplementedError()
 
     def _collect_files(self, tmp_dir: str, **kwargs) -> list: