+++ modelstore/models/pytorch.py
@@ -50,7 +50,7 @@ class PyTorchManager(ModelManager):
         return {"library": "pytorch"}
 
     def _model_data(self, **kwargs) -> dict:
-        """ Returns meta-data about features used to train the model """
+        """ Returns meta-data about the data used to train the model """
         return {}
 
     def _get_functions(self, **kwargs) -> list: