+++ modelstore/models/tensorflow.py
@@ -45,7 +45,7 @@ class TensorflowManager(ModelManager):
         return {"library": "tensorflow"}
 
     def _model_data(self, **kwargs) -> dict:
-        """ Returns meta-data about features used to train the model """
+        """ Returns meta-data about the data used to train the model """
         return {}
 
     def _get_functions(self, **kwargs) -> list: