+++ modelstore/models/transformers.py
@@ -48,7 +48,7 @@ class TransformersManager(ModelManager):
         return {"library": "transformers"}
 
     def _model_data(self, **kwargs) -> dict:
-        """ Returns meta-data about features used to train the model """
+        """ Returns meta-data about the data used to train the model """
         return {}
 
     def _get_functions(self, **kwargs) -> list: