+++ modelstore/models/xgboost.py
@@ -50,7 +50,7 @@ class XGBoostManager(ModelManager):
         return {"library": "xgboost", "type": type(kwargs["model"]).__name__}
 
     def _model_data(self, **kwargs) -> dict:
-        """ Returns meta-data about features used to train the model """
+        """ Returns meta-data about the data used to train the model """
         return {}
 
     def _get_functions(self, **kwargs) -> list: