+++ modelstore/meta/metadata.py
@@ -5,10 +5,11 @@ from modelstore.meta import dependencies, revision, runtime
 
 
 def generate_for_model(
+    domain: str,
     model_id: str,
     model_info: dict,
-    domain: str,
     model_params: dict = None,
+    model_data: dict = None,
 ):
     metadata = {
         "domain": domain,
@@ -17,6 +18,8 @@ def generate_for_model(
     }
     if model_params is not None:
         metadata["parameters"] = _remove_nones(model_params)
+    if model_data is not None:
+        metadata["data"] = _remove_nones(model_data)
     return metadata