+++ modelstore/models/tensorflow.py
@@ -40,9 +40,13 @@ class TensorflowManager(ModelManager):
     def _required_kwargs(self):
         return ["model"]
 
-    def model_info(self, **kwargs) -> dict:
+    def _model_info(self, **kwargs) -> dict:
         """ Returns meta-data about the model's type """
         return {"library": "tensorflow"}
+
+    def _model_data(self, **kwargs) -> dict:
+        """ Returns meta-data about the data used to train the model """
+        return {}
 
     def _get_functions(self, **kwargs) -> list:
         model = kwargs["model"]