+++ tests/models/test_lightgbm.py
@@ -45,7 +45,13 @@ def lightgbm_manager():
 
 def test_model_info(lightgbm_manager, lgb_model):
     exp = {"library": "lightgbm", "type": "Booster"}
-    res = lightgbm_manager.model_info(model=lgb_model)
+    res = lightgbm_manager._model_info(model=lgb_model)
+    assert exp == res
+
+
+def test_model_data(lightgbm_manager, lgb_model):
+    exp = {}
+    res = lightgbm_manager._model_data(model=lgb_model)
     assert exp == res